Merge from upstream/master through github (#5)
* Fix unit test that changes file descriptor limit.

We can lower the file descriptor limit but we don't have the permissions
to increase it.  We therefore lower it by only one.

* Implement key synchronization between enclaves.

When scaling enclaves horizontally, it is desirable that all enclaves
can synchronize their key material, to appear to the outside world as a
single enclave.  This commit implements the key synchronization protocol
that's discussed in Section 3.4.5 of the following paper:
https://arxiv.org/pdf/2206.04123.pdf

This commit adds the following files:

* cache.go -- A cache to keep track of nonces.
* keysync_initiator.go -- The initiating enclave's key sync code.
* keysync_responder.go -- The responding enclave's key sync code.
* keysync_shared.go -- Key sync code shared b/w initiator and responder.

This commit fixes the following issue:
https://github.com/brave-experiments/nitriding/issues/14

* Fix misleading comment.

* Revise cache pruning and add test.

* Fix misleading comment.

* Limit how much we're willing to read from body.

* Limit attsttn doc length we're willing to read.

* Upgrade dependencies.

* Use http.MaxBytesReader to limit request body.

io.ReadFull really wasn't the right way to go about this.

* s/using Kubernetes/horizontally/

* Change content type to text/plain.

...because we're sending Base64-encoded data.

* Implement limitReader.

This will help us read a limited amount of bytes from an HTTP response.

* Reorder checks and use Helper for clearer errors.

* Verify PCR values.

* Move from secretbox to box.

* Clean up code and add complex unit test.

* Only return error if more than the limit was read.

* s/box/NaCl box/

* Don't ignore given public key.

Clearly a bug that I stumbled upon by accident.  Yikes.

* Put encrypted material into UserData field.

* Remove Bytes and newBoxKeyFromBytes.

Now that we no longer have a nonce as part of the box key, we can get
rid of some extra code.

* Use io.ReadAll instead of ioutil.ReadAll.

...cause it's deprecated.

* Add function documentation for getKeysHandler.

* Bundle attestation doc info for tests.

Move everything that belongs together in a single place.  Also, add a
unit test that tests the responding enclave's HTTP handler.

* Don't read one byte more than necessary.

* Explain why we're sending an extra byte.

* Convert byte buffer to string and use Fprint.

That results in the client receiving the exact number of bytes that
we're sending.

* Fix two off-by-one bugs in limit reader and test.

* Wrap errors and store them as error type.

The advantage of this is that it becomes easier to check for specific
errors in our unit tests, by using errors.Is().

This fixes issue https://github.com/brave-experiments/nitriding/issues/18.

* Add unit test that verifies EOF behavior.

* Check if all return values match expectations.

* Add codeql-analysis.yml workflow.

As suggested in our GitHub security policy checklist.

Co-authored-by: Philipp Winter <phw@brave.com>